+++ src/main.cc
@@ -19,11 +19,11 @@
 
 int main() {
   std::cout << "<< BIENVENIDO AL PROGRAMA DE CAMBIO DE MONEDAS >>" << std::endl;
-  int amount;
+  float amount;
   std::cout << "Introduzca el valor objetivo, el cual quiere comprobar el número total de monedas: ";
   std::cin >> amount;
   coinsExchange coinsExchangeObject;
-  coinsExchangeObject.setAmount(static_cast<float>(amount));
+  coinsExchangeObject.setAmount(amount);
   coinsExchangeObject.SolutionFunction();
   return 0;
 }
